requirements.txt
tqdm==4.67.1
sklearn-crfsuite==0.5.0
FlagEmbedding==1.3.4
faiss-gpu==1.7.2
numpy==1.26.4
pytrec_eval==0.5
python-dotenv==1.0.1
openai==1.61.1